% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_table_day.R
\name{cohort_table_day}
\alias{cohort_table_day}
\title{Create a Cohort Table Using Day Level Event Data}
\usage{
cohort_table_day(df, id_var, date)
}
\arguments{
\item{df}{Dataframe}

\item{id_var}{ID variable}

\item{date}{Date}
}
\value{
Cohort table
}
\description{
Creates a cohort table with day level event data with rows corresponding to cohort numbers and columns as dates.
}
\examples{
cohort_table_day(gamelaunch, userid, eventDate)


}
