% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_table_month.R
\name{cohort_table_month}
\alias{cohort_table_month}
\title{Create a Cohort Table Using Month Level Event Data}
\usage{
cohort_table_month(df, id_var, date)
}
\arguments{
\item{df}{Dataframe}

\item{id_var}{ID variable}

\item{date}{Date}
}
\value{
Cohort table
}
\description{
Creates a cohort table with month level event data with rows corresponding to cohort numbers and columns as months.
}
\examples{
cohort_table_month(online_cohorts, CustomerID, InvoiceDate)

}
