% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_table_pct.R
\name{cohort_table_pct}
\alias{cohort_table_pct}
\title{Convert Values of a Cohort Table to Percentages}
\usage{
cohort_table_pct(cohort_table, decimals = 1)
}
\arguments{
\item{cohort_table}{Cohort table}

\item{decimals}{Integer}
}
\value{
Cohort table
}
\description{
Converts values of a cohort table to percentages of initial cohort sizes.
}
\examples{
online_cohorts \%>\%
cohort_table_month(CustomerID, InvoiceDate) \%>\%
cohort_table_pct(decimals = 1)

}
