% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_table_week.R
\name{cohort_table_week}
\alias{cohort_table_week}
\title{Create a Cohort Table Using Week Level Event Data}
\usage{
cohort_table_week(df, id_var, date)
}
\arguments{
\item{df}{Dataframe}

\item{id_var}{ID variable}

\item{date}{Date}
}
\value{
Cohort table
}
\description{
Creates a cohort table with week level event data with rows corresponding to cohort numbers and columns as dates.
}
\examples{
cohort_table_week(gamelaunch, userid, eventDate)


}
