% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_table_year.R
\name{cohort_table_year}
\alias{cohort_table_year}
\title{Create a Cohort Table Using Year Level Event Data}
\usage{
cohort_table_year(df, id_var, date)
}
\arguments{
\item{df}{Dataframe}

\item{id_var}{ID variable}

\item{date}{Date}
}
\value{
Cohort table
}
\description{
Creates a cohort table with year level event data with rows corresponding to cohort numbers and columns as dates.
}
\examples{
cohort_table_week(gamelaunch, userid, eventDate)


}
