% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelaunch.R
\docType{data}
\name{gamelaunch}
\alias{gamelaunch}
\title{Event Data From the Launch of a Mobile Game}
\format{
A data frame with 29324 rows and 3 variables:
\describe{
\item{userid}{User ID}
\item{eventDate}{Event Date}
\item{eventName}{Event Name}
...
}
}
\usage{
gamelaunch
}
\description{
A dataset containing 29324 observations with three columns: userid, eventDate and eventName.
}
\keyword{datasets}
