% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_cohorts.R
\docType{data}
\name{online_cohorts}
\alias{online_cohorts}
\title{Order Data From an Online Store}
\format{
A data frame with 19,573 rows and 2 variables:
\describe{
\item{CustomerID}{Customer ID}
\item{InvoiceDate}{InvoiceDate}
...
}
}
\usage{
online_cohorts
}
\description{
A dataset containing 19573 observations with two columns: CustomerID and InvoiceDate
}
\keyword{datasets}
