% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_left.R
\name{shift_left}
\alias{shift_left}
\title{Left-shift a Cohort Table}
\usage{
shift_left(cohort_table)
}
\arguments{
\item{cohort_table}{Cohort table}
}
\value{
Cohort table
}
\description{
Left-shifts a cohort table
}
\examples{
online_cohorts \%>\%
cohort_table_month(CustomerID, InvoiceDate) \%>\%
shift_left()

}
