% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_left_pct.R
\name{shift_left_pct}
\alias{shift_left_pct}
\title{Left-shift a Cohort Table With Cohort Sizes as Percentages}
\usage{
shift_left_pct(cohort_table, decimals = 1)
}
\arguments{
\item{cohort_table}{Cohort table}

\item{decimals}{Integer}
}
\value{
Cohort table
}
\description{
Left-shifts a cohort table with cohort sizes as percentages of start sizes.
}
\examples{
online_cohorts \%>\%
cohort_table_month(CustomerID, InvoiceDate) \%>\%
shift_left_pct()

}
